Create queclink package and parser shim
