+++ queclink/messages/gteri.py
@@ -1,0 +1,6 @@
+# Analizador de mensajes +RESP:GTERI y variantes Queclink.
+"""Interfaz orientada a mensajes GTERI reutilizando el analizador general."""
+
+from ..parser import parse_gteri
+
+__all__ = ["parse_gteri"]
