+++ queclink/messages/gtinf.py
@@ -1,0 +1,6 @@
+# Analizador de mensajes +RESP:GTINF y variantes de inventario.
+"""Punto de entrada para analizadores de mensajes GTINF (por implementar)."""
+
+from ..parser import parse_gtinf
+
+__all__ = ["parse_gtinf"]
