+++ queclink/specs/__init__.py
@@ -1,0 +1,6 @@
+# Subpaquete para cargar especificaciones de mensajes Queclink.
+"""Herramientas relacionadas con esquemas y metadatos de mensajes."""
+
+from .loader import load_spec
+
+__all__ = ["load_spec"]
