+++ tests/conftest.py
@@ -1,0 +1,10 @@
+"""Configuración de pruebas compartida para facilitar los imports del paquete."""
+
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+ROOT = Path(__file__).resolve().parents[1]
+if str(ROOT) not in sys.path:
+    sys.path.insert(0, str(ROOT))
